+++ R/classdist.R
@@ -1,6 +1,6 @@
 #' Distances to Class Centroids
 #'
-#' `step_classdist` creates a a *specification* of a
+#' `step_classdist` creates a *specification* of a
 #'  recipe step that will convert numeric data into Mahalanobis
 #'  distance measurements to the data centroid. This is done for
 #'  each value of a categorical class variable.
@@ -32,12 +32,10 @@
 #' @keywords datagen
 #' @concept preprocessing dimension_reduction
 #' @export
-#' @details `step_classdist` will create a
-#'
-#' The function will create a new column for every unique value of
-#'  the `class` variable. The resulting variables will not
-#'  replace the original values and have the prefix
-#'  `classdist_`.
+#' @details `step_classdist` will create a new column for every
+#'  unique value of the `class` variable.
+#'  The resulting variables will not replace the original values
+#'  and have the prefix `classdist_`.
 #'
 #' Note that, by default, the default covariance function requires
 #'  that each class should have at least as many rows as variables